% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumersurveys_functions.R
\docType{package}
\name{consumersurveys_googleAuthR}
\alias{consumersurveys_googleAuthR}
\alias{consumersurveys_googleAuthR-package}
\title{Consumer Surveys API
Creates and conducts surveys, lists the surveys that an authenticated user owns, and retrieves survey results and information about specified surveys.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 19:39:58
filename: /Users/mark/dev/R/autoGoogleAPI/googleconsumersurveysv2.auto/R/consumersurveys_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/consumersurveys
\item https://www.googleapis.com/auth/consumersurveys.readonly
\item https://www.googleapis.com/auth/userinfo.email
}
}
