% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumersurveys_objects.R
\name{MobileAppPanel}
\alias{MobileAppPanel}
\title{MobileAppPanel Object}
\usage{
MobileAppPanel(country = NULL, isPublicPanel = NULL, language = NULL,
  mobileAppPanelId = NULL, name = NULL, owners = NULL)
}
\arguments{
\item{country}{Country code for the country of the users that the panel contains}

\item{isPublicPanel}{Whether or not the panel is accessible to all API users}

\item{language}{Language code that the panel can target}

\item{mobileAppPanelId}{Unique panel ID string}

\item{name}{Human readable name of the audience panel}

\item{owners}{List of email addresses for users who can target members of this panel}
}
\value{
MobileAppPanel object
}
\description{
MobileAppPanel Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Representation of an individual pre-defined panel object defining a targeted audience of opinion rewards mobile app users.
}
\seealso{
Other MobileAppPanel functions: \code{\link{mobileapppanels.update}}
}

