% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumersurveys_functions.R
\name{mobileapppanels.update}
\alias{mobileapppanels.update}
\title{Updates a MobileAppPanel. Currently the only property that can be updated is the owners property.}
\usage{
mobileapppanels.update(MobileAppPanel, panelId)
}
\arguments{
\item{MobileAppPanel}{The \link{MobileAppPanel} object to pass to this method}

\item{panelId}{External URL ID for the panel}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/consumersurveys
\item https://www.googleapis.com/auth/userinfo.email
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/consumersurveys, https://www.googleapis.com/auth/userinfo.email)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{}{Google Documentation}

Other MobileAppPanel functions: \code{\link{MobileAppPanel}}
}
