% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumersurveys_objects.R
\name{PageInfo}
\alias{PageInfo}
\title{PageInfo Object}
\usage{
PageInfo(resultPerPage = NULL, startIndex = NULL, totalResults = NULL)
}
\arguments{
\item{resultPerPage}{No description}

\item{startIndex}{No description}

\item{totalResults}{No description}
}
\value{
PageInfo object
}
\description{
PageInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
