% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumersurveys_objects.R
\name{ResultsGetRequest}
\alias{ResultsGetRequest}
\title{ResultsGetRequest Object}
\usage{
ResultsGetRequest(resultMask = NULL)
}
\arguments{
\item{resultMask}{No description}
}
\value{
ResultsGetRequest object
}
\description{
ResultsGetRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other ResultsGetRequest functions: \code{\link{results.get}}
}

