% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumersurveys_objects.R
\name{Survey}
\alias{Survey}
\title{Survey Object}
\usage{
Survey(audience = NULL, cost = NULL, customerData = NULL,
  description = NULL, owners = NULL, questions = NULL,
  rejectionReason = NULL, state = NULL, surveyUrlId = NULL,
  title = NULL, wantedResponseCount = NULL)
}
\arguments{
\item{audience}{No description}

\item{cost}{No description}

\item{customerData}{No description}

\item{description}{No description}

\item{owners}{No description}

\item{questions}{No description}

\item{rejectionReason}{No description}

\item{state}{No description}

\item{surveyUrlId}{No description}

\item{title}{No description}

\item{wantedResponseCount}{No description}
}
\value{
Survey object
}
\description{
Survey Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other Survey functions: \code{\link{surveys.insert}},
  \code{\link{surveys.update}}
}
