% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumersurveys_objects.R
\name{SurveyAudience}
\alias{SurveyAudience}
\title{SurveyAudience Object}
\usage{
SurveyAudience(ages = NULL, country = NULL, countrySubdivision = NULL,
  gender = NULL, languages = NULL, mobileAppPanelId = NULL,
  populationSource = NULL)
}
\arguments{
\item{ages}{No description}

\item{country}{No description}

\item{countrySubdivision}{No description}

\item{gender}{No description}

\item{languages}{No description}

\item{mobileAppPanelId}{No description}

\item{populationSource}{No description}
}
\value{
SurveyAudience object
}
\description{
SurveyAudience Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
