% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumersurveys_objects.R
\name{SurveyCost}
\alias{SurveyCost}
\title{SurveyCost Object}
\usage{
SurveyCost(costPerResponseNanos = NULL, currencyCode = NULL,
  maxCostPerResponseNanos = NULL, nanos = NULL)
}
\arguments{
\item{costPerResponseNanos}{No description}

\item{currencyCode}{No description}

\item{maxCostPerResponseNanos}{No description}

\item{nanos}{No description}
}
\value{
SurveyCost object
}
\description{
SurveyCost Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
