% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumersurveys_objects.R
\name{SurveyQuestion}
\alias{SurveyQuestion}
\title{SurveyQuestion Object}
\usage{
SurveyQuestion(answerOrder = NULL, answers = NULL, hasOther = NULL,
  highValueLabel = NULL, images = NULL, lastAnswerPositionPinned = NULL,
  lowValueLabel = NULL, mustPickSuggestion = NULL, numStars = NULL,
  openTextPlaceholder = NULL, openTextSuggestions = NULL, question = NULL,
  sentimentText = NULL, singleLineResponse = NULL,
  thresholdAnswers = NULL, type = NULL, unitOfMeasurementLabel = NULL,
  videoId = NULL)
}
\arguments{
\item{answerOrder}{No description}

\item{answers}{No description}

\item{hasOther}{No description}

\item{highValueLabel}{No description}

\item{images}{No description}

\item{lastAnswerPositionPinned}{No description}

\item{lowValueLabel}{No description}

\item{mustPickSuggestion}{No description}

\item{numStars}{No description}

\item{openTextPlaceholder}{No description}

\item{openTextSuggestions}{No description}

\item{question}{No description}

\item{sentimentText}{No description}

\item{singleLineResponse}{No description}

\item{thresholdAnswers}{No description}

\item{type}{No description}

\item{unitOfMeasurementLabel}{No description}

\item{videoId}{No description}
}
\value{
SurveyQuestion object
}
\description{
SurveyQuestion Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
