% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumersurveys_objects.R
\name{SurveyQuestionImage}
\alias{SurveyQuestionImage}
\title{SurveyQuestionImage Object}
\usage{
SurveyQuestionImage(altText = NULL, data = NULL, url = NULL)
}
\arguments{
\item{altText}{No description}

\item{data}{No description}

\item{url}{No description}
}
\value{
SurveyQuestionImage object
}
\description{
SurveyQuestionImage Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
