% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumersurveys_objects.R
\name{SurveyRejection}
\alias{SurveyRejection}
\title{SurveyRejection Object}
\usage{
SurveyRejection(explanation = NULL, type = NULL)
}
\arguments{
\item{explanation}{No description}

\item{type}{No description}
}
\value{
SurveyRejection object
}
\description{
SurveyRejection Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
