% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumersurveys_objects.R
\name{SurveyResults}
\alias{SurveyResults}
\title{SurveyResults Object}
\usage{
SurveyResults(status = NULL, surveyUrlId = NULL)
}
\arguments{
\item{status}{No description}

\item{surveyUrlId}{No description}
}
\value{
SurveyResults object
}
\description{
SurveyResults Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
