% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumersurveys_objects.R
\name{SurveysDeleteResponse}
\alias{SurveysDeleteResponse}
\title{SurveysDeleteResponse Object}
\usage{
SurveysDeleteResponse(requestId = NULL)
}
\arguments{
\item{requestId}{Unique request ID used for logging and debugging}
}
\value{
SurveysDeleteResponse object
}
\description{
SurveysDeleteResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
