% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumersurveys_objects.R
\name{SurveysListResponse}
\alias{SurveysListResponse}
\title{SurveysListResponse Object}
\usage{
SurveysListResponse(pageInfo = NULL, requestId = NULL, resources = NULL,
  tokenPagination = NULL)
}
\arguments{
\item{pageInfo}{No description}

\item{requestId}{Unique request ID used for logging and debugging}

\item{resources}{An individual survey resource}

\item{tokenPagination}{No description}
}
\value{
SurveysListResponse object
}
\description{
SurveysListResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
