% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumersurveys_objects.R
\name{SurveysStartResponse}
\alias{SurveysStartResponse}
\title{SurveysStartResponse Object}
\usage{
SurveysStartResponse(requestId = NULL, resource = NULL)
}
\arguments{
\item{requestId}{Unique request ID used for logging and debugging}

\item{resource}{Survey object containing the specification of the started Survey}
}
\value{
SurveysStartResponse object
}
\description{
SurveysStartResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
