% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumersurveys_objects.R
\name{SurveysStopResponse}
\alias{SurveysStopResponse}
\title{SurveysStopResponse Object}
\usage{
SurveysStopResponse(requestId = NULL, resource = NULL)
}
\arguments{
\item{requestId}{Unique request ID used for logging and debugging}

\item{resource}{Survey object containing the specification of the stopped Survey}
}
\value{
SurveysStopResponse object
}
\description{
SurveysStopResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

