% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumersurveys_objects.R
\name{MobileAppPanel}
\alias{MobileAppPanel}
\title{MobileAppPanel Object}
\usage{
MobileAppPanel(country = NULL, isPublicPanel = NULL, language = NULL,
  mobileAppPanelId = NULL, name = NULL, owners = NULL)
}
\arguments{
\item{country}{No description}

\item{isPublicPanel}{No description}

\item{language}{No description}

\item{mobileAppPanelId}{No description}

\item{name}{No description}

\item{owners}{No description}
}
\value{
MobileAppPanel object
}
\description{
MobileAppPanel Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other MobileAppPanel functions: \code{\link{mobileapppanels.update}}
}
