% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumersurveys_objects.R
\name{SurveyCost}
\alias{SurveyCost}
\title{SurveyCost Object}
\usage{
SurveyCost(costPerResponseNanos = NULL, currencyCode = NULL,
  maxCostPerResponseNanos = NULL, nanos = NULL)
}
\arguments{
\item{costPerResponseNanos}{Cost per survey response in nano units of the given currency}

\item{currencyCode}{Currency code that the cost is given in}

\item{maxCostPerResponseNanos}{Threshold to start a survey automatically if the quoted price is at most this value}

\item{nanos}{Cost of survey in nano units of the given currency}
}
\value{
SurveyCost object
}
\description{
SurveyCost Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Message defining the cost to run a given survey through API.
}

