% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumersurveys_objects.R
\name{SurveyQuestion}
\alias{SurveyQuestion}
\title{SurveyQuestion Object}
\usage{
SurveyQuestion(answerOrder = NULL, answers = NULL, hasOther = NULL,
  highValueLabel = NULL, images = NULL, lastAnswerPositionPinned = NULL,
  lowValueLabel = NULL, mustPickSuggestion = NULL, numStars = NULL,
  openTextPlaceholder = NULL, openTextSuggestions = NULL, question = NULL,
  sentimentText = NULL, singleLineResponse = NULL,
  thresholdAnswers = NULL, type = NULL, unitOfMeasurementLabel = NULL,
  videoId = NULL)
}
\arguments{
\item{answerOrder}{The randomization option for multiple choice and multi-select questions}

\item{answers}{Required list of answer options for a question}

\item{hasOther}{Option to allow open-ended text box for Single Answer and Multiple Answer question types}

\item{highValueLabel}{For rating questions, the text for the higher end of the scale, such as 'Best'}

\item{images}{No description}

\item{lastAnswerPositionPinned}{Currently only support pinning an answer option to the last position}

\item{lowValueLabel}{For rating questions, the text for the lower end of the scale, such as 'Worst'}

\item{mustPickSuggestion}{Option to force the user to pick one of the open text suggestions}

\item{numStars}{Number of stars to use for ratings questions}

\item{openTextPlaceholder}{Placeholder text for an open text question}

\item{openTextSuggestions}{A list of suggested answers for open text question auto-complete}

\item{question}{Required question text shown to the respondent}

\item{sentimentText}{Used by the Rating Scale with Text question type}

\item{singleLineResponse}{Option to allow multiple line open text responses instead of a single line response}

\item{thresholdAnswers}{The threshold/screener answer options, which will screen a user into the rest of the survey}

\item{type}{Required field defining the question type}

\item{unitOfMeasurementLabel}{Optional unit of measurement for display (for example: hours, people, miles)}

\item{videoId}{The YouTube video ID to be show in video questions}
}
\value{
SurveyQuestion object
}
\description{
SurveyQuestion Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Message defining the question specifications.
}

