% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumersurveys_objects.R
\name{SurveyQuestionImage}
\alias{SurveyQuestionImage}
\title{SurveyQuestionImage Object}
\usage{
SurveyQuestionImage(altText = NULL, data = NULL, url = NULL)
}
\arguments{
\item{altText}{The alt text property used in image tags is required for all images}

\item{data}{Inline jpeg, gif, tiff, bmp, or png image raw bytes for an image question types}

\item{url}{The read-only URL for the hosted images}
}
\value{
SurveyQuestionImage object
}
\description{
SurveyQuestionImage Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Container object for image data and alt_text.
}

